server/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
boto3
django-cors-headers
python-dotenv
requests
